backend/requirements.txt
# Core web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# AWS and serverless dependencies
boto3==1.34.144
botocore==1.34.144
mangum==0.17.0  # ASGI adapter for AWS Lambda

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Data validation and serialization
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.1.2  # Required for EmailStr validation

# HTTP client for external APIs
httpx==0.27.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production server
gunicorn==21.2.0

# Logging and monitoring
python-json-logger==2.0.7

# Date and time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development and debugging
black==23.11.0  # code formatting
flake8==6.1.0   # linting
mypy==1.7.1     # type checking

# Logging and monitoring
structlog==23.2.0

# JSON handling
orjson==3.9.10
reportlab==4.0.7
